--- a/src/group/group.js
+++ b/src/group/group.js
@@ -2,31 +2,17 @@ import config from '../config/config'
 import { gsap, debug } from '../utils'
 import Timelines from './timelines'
 import { EventEmitter } from 'events'
-
-/**
- * Group defaults
- *
- * @type {object}
- */
-export const groupDefaults = {
-  name: undefined,
-  fps: 30,
-  timelines: new Timelines()
-}
+import { emitChange } from '../utils/emitter'
 
 /**
  * Group.
  */
 class Group extends EventEmitter {
 
-  _name = groupDefaults.name
-  _fps = groupDefaults.fps
-  _timelines = groupDefaults.timelines
+  _name = 'untitled'
+  _timeScale = 1
+  _timelines = new Timelines()
 
-  /**
-   * Gsap timeline
-   * @type {null|TimelineLite|TimelineMax}
-   */
   timeline = null
 
   /**
@@ -42,7 +28,16 @@ class Group extends EventEmitter {
       throw new Error('Cannot create group without a name.')
     }
 
-    Object.assign(this, { ...groupDefaults, ...props })
+    const defaults = {
+      name: 'untitled',
+      timeScale: 1,
+      timelines: new Timelines()
+    }
+
+    Object.assign(this, {
+      ...defaults,
+      ...props
+    })
   }
 
   /**
@@ -59,6 +54,7 @@ class Group extends EventEmitter {
    *
    * @param {Timelines} timelines
    */
+  @emitChange()
   set timelines(timelines) {
     if (!(timelines instanceof Timelines)) {
       timelines = new Timelines(Array.from(timelines))
@@ -67,27 +63,54 @@ class Group extends EventEmitter {
   }
 
   /**
-   * Get current fps
+   * Get current timeScale
    *
    * @returns {number}
    */
-  get fps() {
-    return this._fps
+  get timeScale() {
+    return this._timeScale
   }
 
   /**
    * Set fps
    *
-   * @param {number} fps
+   * @param {number} scale
    */
-  set fps(fps) {
-    if (!(typeof fps === 'number' && isFinite(fps))) {
-      throw new Error('Fps needs to be a number')
+  @emitChange()
+  set timeScale(scale) {
+    if (!(typeof scale === 'number' && isFinite(scale))) {
+      throw new Error('timeScale needs to be a number')
     }
-    this._fps = fps
 
     if (this.timeline && this.timeline instanceof config.gsap.timeline) {
-      this.timeline.timeScale(fps / 60)
+      this.timeline.timeScale(scale)
+    }
+
+    this._timeScale = scale
+  }
+
+  /**
+   * Get the timeline duration.
+   * Equal to this.timeline.duration()
+   *
+   * @returns {number}
+   */
+  get duration() {
+    return this.timeline ? this.timeline.duration() : 0
+  }
+
+  /**
+   * Set the timeline duration.
+   * Updates the group timeScale
+   *
+   * @param {number} val
+   */
+  @emitChange()
+  set duration(val) {
+    if (this.timeline && this.timeline instanceof config.gsap.timeline) {
+      this.timeline.duration(val)
+      this.timeScale = this.timeline.timeScale()
+      this._duration = this.timeline.duration()
     }
   }
 
@@ -105,6 +128,7 @@ class Group extends EventEmitter {
    *
    * @param {string} name
    */
+  @emitChange()
   set name(name) {
     if (typeof name !== 'string') {
       throw new Error('Name needs to be a string')
@@ -118,15 +142,11 @@ class Group extends EventEmitter {
    * @returns {object}
    */
   toObject() {
-    const fps = this.fps
     const name = this.name
+    const timeScale = this.timeScale
     const timelines = this.timelines.toArray()
 
-    return {
-      fps,
-      name,
-      timelines
-    }
+    return { name, timeScale, timelines }
   }
 
   /**
@@ -143,57 +163,55 @@ class Group extends EventEmitter {
             
             Did you forgot to call spirit.setup() ?
             
-            spirit.setup() usage:
-            
-                // auto inject gsap from cdn:
-                spirit.setup()
+            @usage
                 
-                // or provide gsap instances manually:
-                spirit.setup({
-                  tween:    TweenMax,
-                  timeline: TimelineMax
+                spirit.setup().then(function(){
+                  // gsap is loaded here..
                 })
           `)
         }
         throw new Error('GSAP cannot be found')
       }
 
-      // initiate an empty gsap timeline
+      // initiate an empty GSAP timeline
       if (this.timeline && this.timeline instanceof config.gsap.timeline) {
         this.timeline.stop()
         this.timeline.kill()
         this.timeline.clear()
       } else {
-        this.timeline = new config.gsap.timeline({  // eslint-disable-line new-cap
-          useFrames: true,
-          paused: true
-        })
+        this.timeline = new config.gsap.timeline({ paused: true }) // eslint-disable-line new-cap
         this.timeline.autoRemoveChildren = false
       }
 
-      // create a valid gsap timeline out of timelines
-      this.timelines.list.filter(tl => tl.type === 'dom').forEach(tl => {
-        const el = tl.transformObject
+      // create a valid GSAP timeline out of timelines
+      this.timelines.each(timeline => {
+        if (timeline.type === 'dom') {
 
-        if (!(el instanceof window.Element)) {
-          throw new Error('transformObject is not an Element')
+          const el = timeline.transformObject
+
+          if (!(el instanceof window.Element)) {
+            throw new Error('transformObject is not an Element')
+          }
+
+          // kill existing tweens
+          config.gsap.tween.killTweensOf(el)
+          delete el._gsTransform
+          delete el._gsTweenID
+          el.setAttribute('style', '')
+
+          // generate new timelines
+          this.timeline.add(gsap.generateTimeline(timeline).play(), 0, 'start')
         }
-
-        // kill existing tweens
-        config.gsap.tween.killTweensOf(el)
-        delete el._gsTransform
-        delete el._gsTweenID
-        el.setAttribute('style', '')
-
-        // stack timelines to group timeline
-        this.timeline.add(gsap.generateTimeline(tl).play(), 0)
       })
 
-      // update timescale based on fps
-      this.timeline.timeScale(this.fps / 60)
+      // update timescale
+      this.timeline.timeScale(this.timeScale)
+      this._duration = this.timeline.duration()
     } catch (err) {
       throw new Error(`Could not construct timeline: ${err.message}`)
     }
+
+    this.emit('construct', this.timeline)
     return this.timeline
   }