--- a/src/group/keyframes.js
+++ b/src/group/keyframes.js
@@ -52,4 +52,14 @@ class Keyframes extends List {
   }
 }
 
+Keyframes.events = [
+  'change:list',
+  'add',
+  'remove',
+  'change',
+  'change:time',
+  'change:value',
+  'change:ease'
+]
+
 export default Keyframes