--- a/src/group/keyframes.js
+++ b/src/group/keyframes.js
@@ -52,7 +52,7 @@ class Keyframes extends List {
   }
 }
 
-Keyframes.events = [
+Keyframes.Events = [
   'change:list',
   'add',
   'remove',