--- a/src/group/prop.js
+++ b/src/group/prop.js
@@ -1,0 +1,5 @@
+class Prop {
+
+}
+
+export default Prop