--- a/src/group/props.js
+++ b/src/group/props.js
@@ -1,0 +1,5 @@
+class Props {
+
+}
+
+export default Props