… package.json
… src/group/params.js
… src/group/timeline.js
… src/group/transition.js
… src/group/transitions.js
… src/index.js
… src/list/list.js
… src/utils/debug.js
… src/utils/emitter.js
… src/utils/events.js
… src/utils/gsap.js
… src/utils/index.js
… src/utils/is.js
… test/fixtures/group/groups-data.js
… test/fixtures/group/groups.js
… test/group-spec.js
… test/groups-spec.js
… test/keyframe-spec.js
… test/keyframes-spec.js
… test/list-spec.js
… test/param-spec.js
… test/params-spec.js
… test/parser-spec.js
… test/prop-spec.js
… test/props-spec.js
… test/registry-spec.js
… test/timeline-spec.js
… test/transition-spec.js
… test/transitions-spec.js
… test/utils-spec.js