This is a major update how timeline data should be interpreted.

# Tasks

- [x] Update timeline data structure
  - [x] Remove transitions completely
  - [x] Introduce properties (as object literal)
  - [x] Map properties internally as `List`
  - [x] Support `Prop.toObject()`
  - [x] Support `Prop.fromObject()`
- [x] Get rid of frames, make everything time-based
- [x] Add tests

# Extra

- [x] Update create/load to use props instead of transitions
- [x] Support `{ props: { x: 10 } }` as well as `{ props: { x: { value: 10, ease: "Power3.easeOut" } } }`
- [x] Expose `window.spirit` for node bundlers to sync with browser extension
- [x] Add prepack to optimise runtime
- [x] Only show warnings in browser environment

## Before

```json
{
  "timeline": {
    "transitions": [
      {
        "frame": 10,
        "ease": "Linear.easeNone",
        "params": [
          { "x": 10 },
          { "y": 10 }
        ]
      }
    ]
  }
}
```

## After

```json
{
  "timeline": {
    "properties": {
      "x": {
        "0.1s": { "value": 10, "ease": "Linear.easeNone" },
        "0.2s": { "value": 20, "ease": "Linear.easeNone" },
        "0.4s": { "value": 10, "ease": "Expo.easeInOut" }
      },
      "y": {
        "0.1s": { "value": 0,   "ease": "Linear.easeNone" },
        "0.2s": { "value": 200, "ease": "Linear.easeNone" },
        "0.4s": { "value": 0,   "ease": "Expo.easeInOut" }
      }
    }
  }
}
```

We might even add grouped eases in the future, like this:

```json
{
  "timeline": {
    "properties": {
      "x": {
        "0.1s": { "value": 10, "ease": "Linear.easeNone" },
        "0.2s": { "value": 20 },
        "0.4s": { "value": 10 }
      },
      "y": {
        "0.1s": { "value": 0 },
        "0.2s": { "value": 200 },
        "0.4s": { "value": 0 }
      }
    }
  },
  "grouped-eases": {
    "x": { 
      "time": { "0.2s": "0.4s" }, 
      "ease": "Custom", "easePath": "M150 0 L75 200 L225 200 Z" 
    },
    "y": {
      "time": { "0.1s": "0.4s" },
      "ease": "Power3.easeInOut"
    }
  }
}
```